master_requirements.txt
altair==4.1.0
appnope==0.1.0
asn1crypto==1.3.0
astor==0.8.1
astroid==2.4.2
attrs==19.3.0
Automat==20.2.0
azure-common==1.1.25
azure-core==1.6.0
azure-storage-blob==12.3.2
backcall==0.1.0
base58==2.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
boto3==1.12.26
botocore==1.15.26
Brotli==1.0.9
cachetools==4.2.1
cchardet==2.1.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
chart-studio==1.1.0
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
dash==1.19.0
dash-auth==1.4.1
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.1
defusedxml==0.6.0
descartes==1.1.0
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
Fiona==1.8.13.post1
Flask==1.1.2
Flask-Compress==1.8.0
Flask-SeaSurf==0.3.0
future==0.18.2
geopandas==0.7.0
gitdb==4.0.5
GitPython==3.1.12
hyperlink==19.0.0
idna==2.9
ijson==2.6.1
incremental==17.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
json5==0.9.1
jsonlines==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
kiwisolver==1.1.0
lab==5.3
lazy-object-proxy==1.4.3
lesscpy==0.14.0
linear-tsv==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
msrest==0.6.17
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
openpyxl==3.0.3
oscrypto==1.2.0
packaging==20.9
pandas==1.1.0
pandocfilters==1.4.2
parsel==1.5.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
ply==3.11
progressbar2==3.50.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Protego==0.1.16
protobuf==3.14.0
psycopg2==2.8.4
ptyprocess==0.6.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodomex==3.9.8
pydeck==0.5.0
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==2.0.0
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyproj==2.6.0
pyrsistent==0.15.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.3.2
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.7.0
Scrapy==1.8.0
seaborn==0.11.1
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.0
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smmap==3.0.5
snowflake-connector-python==2.2.8
snowflake-sqlalchemy==1.2.3
soupsieve==1.9.5
SQLAlchemy==1.3.15
streamlit==0.75.0
tableschema==1.15.0
tabulator==1.36.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.0.3
traitlets==4.3.3
Twisted==19.10.0
tzlocal==2.1
ua-parser==0.10.0
unicodecsv==0.14.1
urllib3==1.25.8
validators==0.18.2
w3lib==1.21.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zope.interface==4.7.1
